add README.md, slides
