--- develop/LICENSE
@@ -1,1 +1,0 @@
-
